requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
Django==5.0.4
django-crispy-forms==2.1
django-environ==0.11.2
idna==3.7
jwt==1.3.1
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
